+++ CHANGES.md
@@ -11,6 +11,7 @@ this and want to give feedback on API changes.
 - By default, don't keep drawing more points to the end of the route
 - Fix bug that created a new waypoint when clicking (and not dragging) an
   intermediate point
+- Double click to end a route
 
 ## 0.1.13